Fix: apply review - modify errorcode status - change pageResponse
