+++ src/main/java/com/fhsh/daitda/order/application/service/query/OrderQueryServiceImpl.java
@@ -30,7 +30,7 @@ public class OrderQueryServiceImpl implements OrderQueryService{
 	// (2) 자동으로 List 객체를 만들어 주입
 	private final List<OrderStrategy> orderStrategies;
 	@Override
-	public GetOrdersResult getOrders(UUID userId, String authRole, Pageable pageable) {
+	public Slice<Order> getOrders(UUID userId, String authRole, Pageable pageable) {
 
 		OrderStrategy strategy = orderStrategies.stream()
 			.filter(s -> s.equal(authRole))
@@ -38,15 +38,8 @@ public class OrderQueryServiceImpl implements OrderQueryService{
 			.orElseThrow(() -> new IllegalArgumentException("지원하지 않는 권한입니다."));
 
 		Slice<Order> orders = strategy.fetch(userId, pageable);
-		GetOrdersResult ordersResult = new GetOrdersResult(orders.stream().map(order -> {
-			return new GetOrdersResult.OrderResult(order.getSupplierCompanyId(),
-				order.getReceiverCompanyId(),
-				order.getOrdererId(),
-				order.getOrderStatus(),
-				order.getCreatedAt());
-		}).toList());
 
-		return ordersResult;
+		return orders;
 	}
 
 	@Override
