+++ src/main/java/com/fhsh/daitda/order/domain/exception/OrderErrorCode.java
@@ -15,7 +15,7 @@ public enum OrderErrorCode implements ErrorCode {
 
 	//NOT FOUND
 	NOT_FOUND_ORDER(HttpStatus.NOT_FOUND, "NOT_FOUND_ORDER"),
-	NOT_MATCH_ORDERER(HttpStatus.BAD_REQUEST, "NOT_MATCH_ORDERER"),
+	NOT_MATCH_ORDERER(HttpStatus.FORBIDDEN, "NOT_MATCH_ORDERER"),
 	ALREADY_FAILED_ORDER(HttpStatus.BAD_REQUEST, "ALREADY_FAILED_ORDER"),
 	ALREADY_CANCELLED(HttpStatus.BAD_REQUEST, "ALREADY_CANCELLED");
 	private final HttpStatus status;
