+++ src/main/java/com/fhsh/daitda/order/presentation/dto/PageResponse.java
@@ -1,0 +1,14 @@
+package com.fhsh.daitda.order.presentation.dto;
+
+import java.util.List;
+
+public class PageResponse<T> {
+	private boolean hasNext;
+	private List<T> data;
+
+	public PageResponse(boolean hasNext, List<T> data) {
+		this.hasNext = hasNext;
+		this.data = data;
+	}
+
+}
